Tweaks admin and moves errormenow and statistics there
This overrides the admin index.html template to add arbitrary pages that
aren't tied to Django models.

This moves the statistics page and the errormenow pages there. Now we
can enhance the statistics page to add more bits to it.
