+++ standup/manage/templates/admin/statistics.html
@@ -1,0 +1,30 @@
+{% extends "admin/base_site.html" %}
+{% load i18n admin_urls static admin_list %}
+
+{% if not is_popup %}
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+  <a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
+  &rsaquo; Management pages
+  &rsaquo; Site statistics
+</div>
+{% endblock %}
+{% endif %}
+
+{% block coltype %}flex{% endblock %}
+
+{% block content %}
+  <div id="content-main">
+    {% for groupname, groupitems in statsitems.items %}
+      <h2>{{ groupname }}</h2>
+      <table class="stats">
+        {% for key, val in groupitems.items %}
+          <tr>
+            <td class="key">{{ key }}</td>
+            <td class="val">{{ val }}</td>
+          </tr>
+        {% endfor %}
+      </table>
+    {% endfor %}
+  </div>
+{% endblock %}
