+++ standup/manage/tests/test_views.py
@@ -1,0 +1,10 @@
+from django.core.urlresolvers import reverse
+from django.test import TestCase
+
+
+class StatisticsViewTestCase(TestCase):
+    def test_unauthenticated_redirects(self):
+        resp = self.client.get(reverse('manage.statistics'))
+        assert resp.status_code == 302
+
+    # FIXME(willkg): add test for authenticated and staff
